Update documentation about Foundation for Apps and Bower
Foundation for Apps is deprecated; update the documentation accordingly.
(Would it be better to completely delete support for Foundation for
Apps?)

Add a note that `foundation update` should only be used for Bower.  (If
you'd prefer that I instead extend it to call Yarn or NPM, I can, but it
seems better to leave that task to Yarn or NPM.)